create CardClient and delete useless code
