--- lang/en/transactions.php
@@ -1,66 +1,0 @@
-<?php
-
-
-use Idez\Caradhras\Enums\Transaction\TransactionOperation;
-
-return [
-    TransactionOperation::CreditBackoffice->value => 'Credit',
-    TransactionOperation::CreditCustom->value => 'Credit',
-    TransactionOperation::Deposit->value => 'Deposit',
-    TransactionOperation::ChargeReceived->value => 'Charge Received',
-    TransactionOperation::DonationSent->value => 'Donation Sent',
-    TransactionOperation::DonationReceived->value => 'Donation Received',
-    TransactionOperation::FeeBankSlip->value => 'Fee',
-    TransactionOperation::FeeCardReissued->value => 'Fee',
-    TransactionOperation::FeeCustom->value => 'Fee',
-    TransactionOperation::FeeIof->value => 'Fee',
-    TransactionOperation::FeeSubscription->value => 'Fee',
-    TransactionOperation::FeeBankTransfer->value => 'Fee',
-    TransactionOperation::FeeTransfer->value => 'Fee',
-    TransactionOperation::FeeWithdrawal->value => 'Fee',
-    TransactionOperation::FeeWithdrawalInternational->value => 'Fee',
-    TransactionOperation::TransferReceived->value => 'Transfer received',
-    TransactionOperation::TransferSent->value => 'Transfer sent',
-    TransactionOperation::PaymentSent->value => 'Pagamento de Título',
-    TransactionOperation::PixReceived->value => 'Pix received',
-    TransactionOperation::PixSent->value => 'Pix sent',
-    TransactionOperation::PixReversalSent->value => 'Pix Refund',
-    TransactionOperation::PixReversalReceived->value => 'Pix Refund',
-    TransactionOperation::PhoneRecharge->value => 'Phone recharge',
-    TransactionOperation::PurchaseInternational->value => 'Purchase',
-    TransactionOperation::PurchaseNational->value => 'Purchase',
-    TransactionOperation::RefundIof->value => 'IOF Refund',
-    TransactionOperation::RefundPayment->value => 'Payment Refund',
-    TransactionOperation::RefundPhoneRecharge->value => 'Phone Recharge Refund',
-    TransactionOperation::RefundPurchaseInternational->value => 'Purchase Refund',
-    TransactionOperation::RefundBankTransfer->value => 'Bank Transfer Refund',
-    TransactionOperation::RefundTransferSent->value => 'Transfer Refund Sent',
-    TransactionOperation::RefundTransferReceived->value => 'Transfer Refund Received',
-    TransactionOperation::RefundPix->value => 'Pix Refund',
-    TransactionOperation::BankTransferReceived->value => 'Transfer received',
-    TransactionOperation::BankTransferSent->value => 'Transfer sent',
-    TransactionOperation::InvoicePaid->value => 'Invoice Paid',
-    TransactionOperation::InvoiceReceived->value => 'Invoice Received',
-    TransactionOperation::InstallmentPaid->value => 'Installment Paid',
-    TransactionOperation::CashbackSent->value => 'Cashback Sent',
-    TransactionOperation::CashbackReceived->value => 'Cashback Recebido',
-    TransactionOperation::CommissionPaid->value => 'Cashback Received',
-    TransactionOperation::CommissionReceived->value => 'Commission Paid',
-    TransactionOperation::BucketWithdrawalSent->value => 'Bucket withdrawal sent',
-    TransactionOperation::BucketWithdrawalReceived->value => 'Bucket withdrawal received',
-    TransactionOperation::BucketCashbackWithdrawalSent->value => 'Cashback withdrawal sent',
-    TransactionOperation::BucketCashbackWithdrawalReceived->value => 'Cashback withdrawal received',
-    TransactionOperation::QrPaymentSale->value => 'Sale with QR Code',
-    TransactionOperation::QrPaymentPurchase->value => 'Purchase with QR Code',
-    TransactionOperation::Withdrawal->value => 'Withdrawal',
-    TransactionOperation::DailyIncome->value => 'Income',
-    TransactionOperation::Other->value => 'Other',
-    TransactionOperation::BacenjudLock->value => 'Impediment by Bacenjud',
-    TransactionOperation::BacenjudTransfer->value => 'Bacenjud Transfer',
-    TransactionOperation::BacenjudUnlock->value => 'Bacenjud Unlock',
-
-    'approved' => 'Transaction approved',
-    'rejected' => 'Transaction rejected',
-    'canceled' => 'Transaction canceled',
-    'expired' => 'Transaction expired',
-];
