--- src/Contracts/NovaSelectEnumInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Idez\Caradhras\Contracts;
-
-interface NovaSelectEnumInterface
-{
-    public function label(): string;
-
-    public static function novaOptions(): array;
-}
