+++ src/Exceptions/TransferFailedException.php
@@ -6,7 +6,7 @@ class TransferFailedException extends CaradhrasException
 {
     public function __construct(string $originalError)
     {
-        parent::__construct(trans('errors.transfer.failed'), 502, 'transfer.failed', [
+        parent::__construct(trans('errors.transfer_failed'), 502, 'transfer_failed', [
             'original_error' => $originalError,
         ]);
     }
